third_party/WebKit/LayoutTests/fast/css-intrinsic-dimensions/max-height-fill-available.html

<!DOCTYPE html>
<style>
  #container { width:fit-content; border:10px solid green; background:red; }
  #child { display:flow-root; border:10px solid green; max-height:-webkit-fill-available; background:red; }
</style>
<p>There should be a green square below, and no red.</p>
<div id="container" data-expected-width="100" data-expected-height="100">
  <div id="child" data-expected-width="80" data-expected-height="80">
    <div style="float:left; width:30px; height:60px; background:green;"></div>
    <div style="float:left; width:30px; height:60px; background:green;"></div>
  </div>
</div>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<script>
  checkLayout("#container");
</script>
